+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v2.1.2
+- added `stubUsersSegments` and `stubAccountsSegments` for separate segments definition
+- marked `stubSegments` as deprecated
+
 # v2.1.1
 - added documentation
 - renamed `getOrgAddr` into `_getOrgAddr` to make it an utility method